Modelos y tablas

+++ database/migrations/2022_11_09_180939_create_productos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('productos', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('codigo')->unique();
+            $table->decimal('codigo',10,0)->unique();
             $table->string('denominación');
             $table->string('descripción');
             $table->decimal('precio',8,2);
